// sniperbot/src/types.ts
export interface WebSocketRequest {
  jsonrpc: string;
  id: number;
  method: string;
  params: unknown[];
}
// ...
export interface CurrentTransaction {
  token: string;
  tokenTime: number;
  purchasePrice: number;
  balance: number;
  solPaid: number;
  solPaidUSDC: number;
  baseVault?: string;
  quoteVault?: string;
  poolAddress?: string;
}